[linkTap] PR Adjustments 1 pass 2
[linkTap] PR Adjustments 1 pass 2
