+++ bundles/org.openhab.binding.linktap/src/main/java/org/openhab/binding/linktap/internal/LinkTapBridgeHandler.java
@@ -123,57 +123,11 @@ public class LinkTapBridgeHandler extends BaseBridgeHandler {
     private final DiscoveryServiceRegistry discoverySrvReg;
 
     @Activate
-    public LinkTapBridgeHandler(final Bridge bridge, @NotNull IHttpClientProvider httpClientProvider,
+    public LinkTapBridgeHandler(final Bridge bridge, IHttpClientProvider httpClientProvider,
             @Reference DiscoveryServiceRegistry discoveryService) {
         super(bridge);
         this.httpClientProvider = httpClientProvider;
         this.discoverySrvReg = discoveryService;
-    }
-
-    public static class Firmware {
-        String raw;
-        int buildVer;
-        int hwVer;
-
-        public Firmware(final @Nullable String fwVersion) {
-            raw = "S00000000";
-            hwVer = 0;
-            buildVer = 0;
-
-            if (fwVersion == null || fwVersion.length() < 7) {
-                return;
-            } else {
-                raw = fwVersion;
-                buildVer = Integer.parseInt(raw.substring(1, 7));
-            }
-
-            switch (fwVersion.charAt(0)) {
-                case 'G':
-                    hwVer = 2;
-                    break;
-                case 'S':
-                    hwVer = 1;
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        public boolean supportsLocalConfig() {
-            return buildVer >= 60883;
-        }
-
-        public boolean supportsMDNS() {
-            return buildVer >= 60880;
-        }
-
-        public String generateTestedRevisionForHw(final int versionNo) {
-            return String.format("%c%05d", raw.charAt(0), versionNo);
-        }
-
-        public String getRecommendedMinVer() {
-            return generateTestedRevisionForHw(60883);
-        }
     }
 
     private void startGwPolling() {
